Refactor components code
